+++ examples/with-auth/src/core/index.ts
@@ -1,12 +1,11 @@
 import { domain } from '@palmares/core';
 import { Response, path, serverDomainModifier } from '@palmares/server';
+import { getAdapters } from 'packages/auth/dist/src/conf';
 
 export default domain('core', import.meta.dirname, {
   modifiers: [serverDomainModifier],
   getRoutes: () =>
     path('/test').get(async () => {
-      return Response.json({
-        message: 'Hello world'
-      });
+     return getAdapters();
     })
 });